+++ src/test/java/com/atlassian/db/replica/api/mocks/PermanentConsistency.java
@@ -3,6 +3,7 @@ package com.atlassian.db.replica.api.mocks;
 import com.atlassian.db.replica.spi.*;
 
 import java.sql.*;
+import java.util.function.Supplier;
 
 public class PermanentConsistency implements ReplicaConsistency {
 
@@ -12,7 +13,7 @@ public class PermanentConsistency implements ReplicaConsistency {
     }
 
     @Override
-    public boolean isConsistent(Connection replica) {
+    public boolean isConsistent(Supplier<Connection> replica) {
         return true;
     }
 }